EPAD8-2151 delete old, start new, each statement
