+++ services/drupal/web/themes/epa_theme/epa_theme.libraries.yml
@@ -10,6 +10,7 @@ global:
       css-lib/colorbox.min.css: {}
   js:
     js/dist/scripts.min.js: {}
+    js/dist/epa-alerts.min.js: {}
     js/dist/sitewide-alert.min.js: {}
     js/dist/definition.min.js: {}
     js/dist/external-links.min.js: {}
